refactor: confeti 관련 파일 및 경로를 performance로 변경
